[new struct of files] create engine folder
Edit Makefile

+++ src/engine/s21_mesh.h
@@ -1,13 +1,12 @@
 #ifndef SRC_S21_MESH_H_
 #define SRC_S21_MESH_H_
 
-#include <stdlib.h>
-#include <stdio.h>
+#include "s21_engine.h"
 
 typedef struct Vec3D {
-    double x;
-    double y;
-    double z;
+    float x;
+    float y;
+    float z;
 } vec3D;
 
 typedef struct Polygon {
@@ -17,6 +16,7 @@ typedef struct Polygon {
 typedef struct Mesh {
     Polygon *polygons;
 } Mesh;
+
 
 Mesh plane();
